+++ modules/ChamCong-backend/ChamCong-backend-service/src/main/resources/META-INF/sql/tables.sql
@@ -62,5 +62,5 @@ create table _Users (
 	created_at DATE null,
 	updated_at DATE null,
 	groupId LONG,
-	userId INTEGER
+	userId LONG
 );
